Fixed a wrong return type. Added new utility methods to directly get the next specific primitive variable. Updated unit tests to include the new methods.

+++ src/main/java/de/jackwhite20/cope/config/Value.java
@@ -84,7 +84,7 @@ public class Value {
      *
      * @return The value as a float.
      */
-    public double asFloat() {
+    public float asFloat() {
 
         return Float.parseFloat(value);
     }
